MDEV-742: read-only, pure myisam, binlog-*, @@skip_log_bin corner cases
Are addressed along the following policies.
Prepared read-only, or on non-transactional engines, or
binlog-* filter in binlog, or skipped to binlog XA:s remains in the xid cache after disconnect.
But their consequent completion with Commit or Rollback differs.

The read-only at reconnect marks XID to fail for future
completion with ER_XA_RBROLLBACK.
The same applies to the non-transactional engine XA.
`binlog-*` filtered XA when it changes engine data is regarded
as loggable even when nothing got cached for binlog.
An empty XA-prepare group is recorded. Consequent Commit-or-Rollback
succeeds in the Engine(s) as well as recorded into binlog.
@@skip_log_bin=OFF does not record anything at XA-prepare (obviously),
but the completion event is recorded into binlog to admit
inconsistency with slave.

The following actions are taken by the patch.

At XA-prepare:
  when empty binlog cache - don't do anything to binlog if RO,
  otherwise write empty XA_prepare (assert(binlog-filter case)).
At Disconnect:
  when Prepared && RO (=> no binlogging was done)
    set Xid_cache_element::error := ER_XA_rollback_by resource (which does
    already take place in Innodb), and
    *keep* XID in the cache (so trans_xa_detach)
At XA-"complete":
   Discover the error, don't binlog the "complete", return the error to the user.
Summary:
 - no binlogging for RO (incl SELECT, zero-row-affected updates, MyISAM) at all
 - binlog-filtered will be logged with empty XA-prepare.